Revert "Change command package to be build package"
