+++ cmd/gh/main.go
@@ -12,7 +12,7 @@ import (
 
 	surveyCore "github.com/AlecAivazis/survey/v2/core"
 	"github.com/cli/cli/api"
-	"github.com/cli/cli/internal/build"
+	"github.com/cli/cli/command"
 	"github.com/cli/cli/internal/config"
 	"github.com/cli/cli/internal/ghinstance"
 	"github.com/cli/cli/internal/run"
@@ -29,12 +29,10 @@ import (
 var updaterEnabled = ""
 
 func main() {
-	buildDate := build.Date
-	buildVersion := build.Version
-
+	currentVersion := command.Version
 	updateMessageChan := make(chan *update.ReleaseInfo)
 	go func() {
-		rel, _ := checkForUpdate(buildVersion)
+		rel, _ := checkForUpdate(currentVersion)
 		updateMessageChan <- rel
 	}()
 
@@ -44,7 +42,7 @@ func main() {
 		ghinstance.OverrideDefault(hostFromEnv)
 	}
 
-	cmdFactory := factory.New(buildVersion)
+	cmdFactory := factory.New(command.Version)
 	stderr := cmdFactory.IOStreams.ErrOut
 	if !cmdFactory.IOStreams.ColorEnabled() {
 		surveyCore.DisableColor = true
@@ -63,7 +61,7 @@ func main() {
 		}
 	}
 
-	rootCmd := root.NewCmdRoot(cmdFactory, buildVersion, buildDate)
+	rootCmd := root.NewCmdRoot(cmdFactory, command.Version, command.BuildDate)
 
 	cfg, err := cmdFactory.Config()
 	if err != nil {
@@ -153,7 +151,7 @@ func main() {
 	if newRelease != nil {
 		msg := fmt.Sprintf("%s %s → %s\n%s",
 			ansi.Color("A new release of gh is available:", "yellow"),
-			ansi.Color(buildVersion, "cyan"),
+			ansi.Color(currentVersion, "cyan"),
 			ansi.Color(newRelease.Version, "cyan"),
 			ansi.Color(newRelease.URL, "yellow"))
 
@@ -213,7 +211,7 @@ func checkForUpdate(currentVersion string) (*update.ReleaseInfo, error) {
 		return nil, nil
 	}
 
-	client, err := basicClient(currentVersion)
+	client, err := command.BasicClient()
 	if err != nil {
 		return nil, err
 	}
@@ -221,31 +219,4 @@ func checkForUpdate(currentVersion string) (*update.ReleaseInfo, error) {
 	repo := updaterEnabled
 	stateFilePath := path.Join(config.ConfigDir(), "state.yml")
 	return update.CheckForUpdate(client, stateFilePath, repo, currentVersion)
-}
-
-// BasicClient returns an API client for github.com only that borrows from but
-// does not depend on user configuration
-func basicClient(currentVersion string) (*api.Client, error) {
-	var opts []api.ClientOption
-	if verbose := os.Getenv("DEBUG"); verbose != "" {
-		opts = append(opts, apiVerboseLog())
-	}
-	opts = append(opts, api.AddHeader("User-Agent", fmt.Sprintf("GitHub CLI %s", currentVersion)))
-
-	token := os.Getenv("GITHUB_TOKEN")
-	if token == "" {
-		if c, err := config.ParseDefaultConfig(); err == nil {
-			token, _ = c.Get(ghinstance.Default(), "oauth_token")
-		}
-	}
-	if token != "" {
-		opts = append(opts, api.AddHeader("Authorization", fmt.Sprintf("token %s", token)))
-	}
-	return api.NewClient(opts...), nil
-}
-
-func apiVerboseLog() api.ClientOption {
-	logTraffic := strings.Contains(os.Getenv("DEBUG"), "api")
-	colorize := utils.IsTerminal(os.Stderr)
-	return api.VerboseLog(utils.NewColorable(os.Stderr), logTraffic, colorize)
 }
