+++ command/root.go
@@ -1,0 +1,54 @@
+package command
+
+import (
+	"fmt"
+	"os"
+	"runtime/debug"
+	"strings"
+
+	"github.com/cli/cli/api"
+	"github.com/cli/cli/internal/config"
+	"github.com/cli/cli/internal/ghinstance"
+	"github.com/cli/cli/utils"
+)
+
+// Version is dynamically set by the toolchain or overridden by the Makefile.
+var Version = "DEV"
+
+// BuildDate is dynamically set at build time in the Makefile.
+var BuildDate = "" // YYYY-MM-DD
+
+func init() {
+	if Version == "DEV" {
+		if info, ok := debug.ReadBuildInfo(); ok && info.Main.Version != "(devel)" {
+			Version = info.Main.Version
+		}
+	}
+}
+
+// BasicClient returns an API client for github.com only that borrows from but
+// does not depend on user configuration
+func BasicClient() (*api.Client, error) {
+	var opts []api.ClientOption
+	if verbose := os.Getenv("DEBUG"); verbose != "" {
+		opts = append(opts, apiVerboseLog())
+	}
+	opts = append(opts, api.AddHeader("User-Agent", fmt.Sprintf("GitHub CLI %s", Version)))
+
+	token := os.Getenv("GITHUB_TOKEN")
+	if token == "" {
+		if c, err := config.ParseDefaultConfig(); err == nil {
+			token, _ = c.Get(ghinstance.Default(), "oauth_token")
+		}
+	}
+	if token != "" {
+		opts = append(opts, api.AddHeader("Authorization", fmt.Sprintf("token %s", token)))
+	}
+	return api.NewClient(opts...), nil
+}
+
+func apiVerboseLog() api.ClientOption {
+	logTraffic := strings.Contains(os.Getenv("DEBUG"), "api")
+	colorize := utils.IsTerminal(os.Stderr)
+	return api.VerboseLog(utils.NewColorable(os.Stderr), logTraffic, colorize)
+}
